Move remaining content of gc_worker/lib.rs to new gc_worker.rs
